requirements.txt
PySide6
insightface
onnxruntime
scikit-learn
numpy
opencv-python
imagehash
scipy
requests
pillow
